catch test files up to renaming i forgot existed, thanks brooke!
