+++ packages/story-editor/src/components/publishModal/index.js
@@ -14,5 +14,3 @@
  * limitations under the License.
  */
 export { default as PublishModal } from './publishModal';
-
-// This whole directory of components in /publishModal is a work in progress!
